+++ src/app/clarity/descriptions/ClarityDescriptions.tsx
@@ -1,12 +1,9 @@
 import ExternalLink from 'app/dim-ui/ExternalLink';
 import { t } from 'app/i18next-t';
 import clsx from 'clsx';
-import React from 'react';
-import { useSelector } from 'react-redux';
-import { descriptionsSelector } from '../selectors';
 /* eslint-disable css-modules/no-unused-class */
 import styles from './Description.m.scss';
-import { LinesContent } from './descriptionInterface';
+import { LinesContent, Perk } from './descriptionInterface';
 
 const customContent = (content: LinesContent) => {
   if (content.linkUrl) {
@@ -20,39 +17,65 @@ const joinClassNames = (classNames?: string) =>
     .map((className) => styles[className])
     .join(' ');
 
-/**
- * @param Object.hash Perk hash from DestinyInventoryItemDefinition
- * @param Object.defaultDimDescription It will return whatever you give it if it can't find the perk
- ** This is cut down version of original converted
- */
-export default function ClarityDescriptions({
-  hash,
-  fallback,
-  communityOnly,
-}: {
-  hash: number;
-  fallback?: React.ReactNode;
-  communityOnly?: boolean;
-}) {
-  const descriptions = useSelector(descriptionsSelector);
-  const lines = descriptions?.[hash]?.simpleDescription;
-  const statOnly = descriptions?.[hash]?.statOnly;
-  if (!lines || statOnly) {
-    return <>{fallback ?? null}</>;
+/*
+       (^|\b) : start from the beginning of the string or a word boundary
+        [+-]? : include + or - prefixes
+  (\d*\.)?\d+ : match numbers (including decimal values)
+([xs]|ms|HP)? : optionally include 'x' multiplier, 's', 'ms' and 'HP' unit suffixes
+      ?:[%°+] : optionally include %, ° and + suffixes
+         \b|$ : stop at a word boundary or the end of the string
+*/
+const boldTextRegEx = /(^|\b)([+-]?(\d*\.)?\d+([xs]|ms|HP)?)(?:[%°+]|\b|$)/g;
+
+function applyFormatting(text: string) {
+  const segments = [];
+
+  const matches = [...text.matchAll(boldTextRegEx)];
+  let startIndex = 0;
+  let n = 0;
+  for (const match of matches) {
+    if (match.index === undefined) {
+      continue;
+    }
+    const capturedText = match[0];
+    segments.push(text.substring(startIndex, match.index));
+    segments.push(<b key={n++}>{capturedText}</b>);
+    startIndex = match.index + capturedText.length;
+  }
+  if (startIndex < text.length) {
+    segments.push(text.substring(startIndex));
   }
 
-  const convertedDescription = lines?.map((line, i) => (
+  return segments;
+}
+
+/**
+ * Renders the Clarity description for the provided Community Insight.
+ * This is a cut-down version of the original from the Clarity extension.
+ */
+export default function ClarityDescriptions({
+  perk,
+  className,
+}: {
+  perk: Perk;
+  className?: string;
+}) {
+  if (!perk.simpleDescription) {
+    return null;
+  }
+
+  const convertedDescription = perk.simpleDescription.map((line, i) => (
     <div className={joinClassNames(line.className)} key={i}>
       {line.lineText?.map((linesContent, i) => (
         <span className={joinClassNames(linesContent.className)} title={linesContent.title} key={i}>
-          {linesContent.text || customContent(linesContent)}
+          {linesContent.text ? applyFormatting(linesContent.text) : customContent(linesContent)}
         </span>
       ))}
     </div>
   ));
 
   return (
-    <div className={clsx(styles.communityDescription, communityOnly && styles.communityOnly)}>
+    <div className={clsx(styles.communityDescription, className)}>
       <h3>{t('MovePopup.CommunityData')}</h3>
       {convertedDescription}
     </div>